Replace the genTool RPC with an AGENT_TOOL node builder
- LS: add AgentToolBuilder, register NodeKind.AGENT_TOOL, and route agents.bal
  through it. Remove the genTool RPC (AgentsGenerator.genTool and helpers, the
  AgentsManagerService handler, GenToolRequest, GenToolsTest). genAgentTool is
  left intact.
- Frontend: add buildAgentToolNode and swap tool-creation callers to
  getSourceCode(buildAgentToolNode(...)), dropping the genTool RPC chain.
