+++ packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/NodeBuilder.java
@@ -30,6 +30,7 @@ import io.ballerina.flowmodelgenerator.core.model.node.AgentBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AgentCallBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AgentIdAuthConfigBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AgentRunBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.AgentToolBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AgentTypeBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AssignBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.AutomationBuilder;
@@ -176,6 +177,7 @@ public abstract class NodeBuilder implements DiagnosticHandler.DiagnosticCapable
         put(NodeKind.AGENT_TYPE, AgentTypeBuilder::new);
         put(NodeKind.AGENT_CALL, AgentCallBuilder::new);
         put(NodeKind.AGENT_RUN, AgentRunBuilder::new);
+        put(NodeKind.AGENT_TOOL, AgentToolBuilder::new);
         put(NodeKind.CLASS_INIT, ClassInitBuilder::new);
         put(NodeKind.MEMORY, MemoryBuilder::new);
         put(NodeKind.SHORT_TERM_MEMORY_STORE, ShortTermMemoryStoreBuilder::new);
