+++ packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/NodeKind.java
@@ -85,6 +85,7 @@ public enum NodeKind {
     AGENT_TYPE,
     AGENT_CALL,
     AGENT_RUN,
+    AGENT_TOOL,
     CLASS_INIT,
     AGENTS,
