+++ packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-ls-extension/src/main/java/io/ballerina/flowmodelgenerator/extension/AgentsManagerService.java
@@ -26,7 +26,6 @@ import io.ballerina.flowmodelgenerator.core.AgentsGenerator;
 import io.ballerina.flowmodelgenerator.core.McpClient;
 import io.ballerina.flowmodelgenerator.extension.request.AddAgentChatServiceRequest;
 import io.ballerina.flowmodelgenerator.extension.request.GenAgentToolRequest;
-import io.ballerina.flowmodelgenerator.extension.request.GenToolRequest;
 import io.ballerina.flowmodelgenerator.extension.request.GetAiModuleOrgRequest;
 import io.ballerina.flowmodelgenerator.extension.request.GetAllAgentsRequest;
 import io.ballerina.flowmodelgenerator.extension.request.GetAllMemoryManagersRequest;
@@ -290,29 +289,6 @@ public class AgentsManagerService implements ExtendedLanguageServerService {
                 response.setError(new RuntimeException("Failed to get MCP tools: " + errorMsg, e));
                 return response;
             }
-        });
-    }
-
-    @JsonRequest
-    public CompletableFuture<GenToolResponse> genTool(GenToolRequest request) {
-        return CompletableFuture.supplyAsync(() -> {
-            GenToolResponse response = new GenToolResponse();
-            try {
-                Path filePath = Path.of(request.filePath());
-                this.workspaceManager.loadProject(filePath);
-                Optional<SemanticModel> semanticModel = this.workspaceManager.semanticModel(filePath);
-                if (semanticModel.isEmpty()) {
-                    return response;
-                }
-
-                AgentsGenerator agentsGenerator = new AgentsGenerator(semanticModel.get());
-                response.setTextEdits(agentsGenerator.genTool(request.flowNode(), request.toolName(),
-                        request.toolParameters(), request.connection(), request.description(), filePath,
-                        this.workspaceManager));
-            } catch (Throwable e) {
-                throw new RuntimeException(e);
-            }
-            return response;
         });
     }
