--- packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-ls-extension/src/test/java/io/ballerina/flowmodelgenerator/extension/agentsmanager/GenToolsTest.java
@@ -1,152 +1,0 @@
-/*
- *  Copyright (c) 2025, WSO2 LLC. (http://www.wso2.com)
- *
- *  WSO2 LLC. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package io.ballerina.flowmodelgenerator.extension.agentsmanager;
-
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.reflect.TypeToken;
-import io.ballerina.flowmodelgenerator.extension.request.GenToolRequest;
-import io.ballerina.modelgenerator.commons.AbstractLSTest;
-import org.eclipse.lsp4j.TextEdit;
-import org.testng.Assert;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * Tests for generate tools.
- *
- * @since 1.0.0
- */
-public class GenToolsTest extends AbstractLSTest {
-
-    private static final Type textEditListType = new TypeToken<Map<String, List<TextEdit>>>() {
-    }.getType();
-
-    @DataProvider(name = "data-provider")
-    @Override
-    protected Object[] getConfigsList() {
-        return new Object[][]{
-                {Path.of("function_tool.json")},
-                {Path.of("remote_action_tool.json")},
-                {Path.of("remote_action_tool2.json")},
-                {Path.of("remote_action_tool3.json")},
-                {Path.of("remote_action_tool4.json")},
-                {Path.of("remote_action_tool5.json")},
-                {Path.of("remote_action_tool6.json")},
-                {Path.of("remote_action_tool7.json")},
-        };
-    }
-
-    @Override
-    @Test(dataProvider = "data-provider")
-    public void test(Path config) throws IOException {
-        Path configJsonPath = configDir.resolve(config);
-        TestConfig testConfig = gson.fromJson(Files.newBufferedReader(configJsonPath), TestConfig.class);
-
-        String filePath =
-                testConfig.source() == null ? "" : sourceDir.resolve(testConfig.source()).toAbsolutePath().toString();
-        GenToolRequest request = new GenToolRequest(filePath, testConfig.diagram(), testConfig.name(),
-                testConfig.toolParameters(), testConfig.toolDescription(), testConfig.connection());
-        JsonObject jsonMap = getResponseAndCloseFile(request, testConfig.source()).getAsJsonObject("textEdits");
-
-        Map<String, List<TextEdit>> actualTextEdits = gson.fromJson(jsonMap, textEditListType);
-
-        boolean assertFailure = false;
-
-        if (actualTextEdits.size() != testConfig.output().size()) {
-            log.info("The number of text edits does not match the expected output.");
-            assertFailure = true;
-        }
-
-        Map<String, List<TextEdit>> newMap = new HashMap<>();
-        for (Map.Entry<String, List<TextEdit>> entry : actualTextEdits.entrySet()) {
-            Path fullPath = Paths.get(entry.getKey());
-            String relativePath = sourceDir.relativize(fullPath).toString();
-
-            List<TextEdit> textEdits = testConfig.output().get(relativePath.replace("\\", "/"));
-            if (textEdits == null) {
-                log.info("No text edits found for the file: " + relativePath);
-                assertFailure = true;
-            } else if (!assertArray("text edits", entry.getValue(), textEdits)) {
-                assertFailure = true;
-            }
-
-            newMap.put(relativePath, entry.getValue());
-        }
-
-        if (assertFailure) {
-            TestConfig updatedConfig =
-                    new TestConfig(testConfig.source(), testConfig.name(), testConfig.toolParameters(),
-                            testConfig.connection(), testConfig.description(), testConfig.toolDescription(),
-                            testConfig.diagram(), newMap);
-//            updateConfig(configJsonPath, updatedConfig);
-            Assert.fail(String.format("Failed test: '%s' (%s)", testConfig.description(), configJsonPath));
-        }
-    }
-
-    @Override
-    protected String getResourceDir() {
-        return "agents_manager";
-    }
-
-    @Override
-    protected Class<? extends AbstractLSTest> clazz() {
-        return GenToolsTest.class;
-    }
-
-    @Override
-    protected String getApiName() {
-        return "genTool";
-    }
-
-    @Override
-    protected String getServiceName() {
-        return "agentManager";
-    }
-
-    /**
-     * Represents the test configuration for the flow model getNodeTemplate API.
-     *
-     * @param source          The source file path
-     * @param name            The name of the tool
-     * @param toolParameters  The parameters for the tool
-     * @param connection      The name of the connection
-     * @param description     The description of the test
-     * @param toolDescription The description of the tool
-     * @param diagram         The flow node diagram
-     * @param output          The expected output
-     */
-    private record TestConfig(String source, String name, JsonElement toolParameters, String connection,
-                              String description, String toolDescription, JsonElement diagram, Map<String,
-            List<TextEdit>> output) {
-
-        public String description() {
-            return description == null ? "" : description;
-        }
-    }
-}
