+++ packages/ballerina-language-server/flow-model-generator/modules/flow-model-generator-ls-extension/src/test/resources/testng.xml
@@ -107,7 +107,6 @@ under the License.
             <!--agents manager-->
             <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.NodeTemplateTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.SourceGeneratorTest"/>
-            <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.GenToolsTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.GetAgentOrgTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.GetAllAgentsTest"/>
             <class name="io.ballerina.flowmodelgenerator.extension.agentsmanager.GetAllModelsTest"/>
